refactor(core): split merge_execution.py into core/merge/ package
Pure relocation of the 1386-LOC keystone into a cohesive core/merge/ package:
errors, ci_rollup (transport resolution + the fetch_* delegators + CI/pipeline
verdict classification), pr_slug_resolution (repo/slug + host-kind + cross-repo
probe), authorization (§17.4.3 identity/substrate guards + MergePrecheck), and
execution (preconditions orchestration, bound merge, post hook, merge_ticket_pr).

The §1993 cut: _code_host_for + fetch_* live in ci_rollup (the lowest layer) so
pr_slug_resolution and execution both depend DOWN on it — no intra-package cycle
under forbid_circular_dependencies. The PR4 seam is preserved verbatim: transport
resolves only via get_backend_provider(), core never imports teatree.backends.

__init__ is the package facade re-exporting the cross-package public surface;
mock.patch targets name the defining submodule, not the facade. Symbol bodies and
error f-strings are byte-for-byte unchanged; execution's module docstring is
condensed (the §17.4 flow lives in BLUEPRINT + the function docstrings) to clear
the 500-LOC module-health cap.

Campaign: architecture-perfect PR5 (god-module splits), merge concern.
