+++ docs/blueprint/factory-architecture.md
@@ -84,7 +84,7 @@ Before executing any merge the loop **must** perform all of the following checks
 
 0. **Branch-currency (conflict-only)** ([#940](https://github.com/souliane/teatree/issues/940)) — the feature branch must *merge cleanly* into the target; being merely behind is **not** a blocker (it squash-merges without a rebase, since GitHub re-applies the diff onto the live target at merge time and steps 2–3 still guard correctness). At **ship** time `branch_currency.require_current_branch` auto-merges a zero-conflict gap (so step 1's attestation binds to the post-merge SHA) and refuses only on conflict; at **CLEAR** time `sha_conflicts_with_target` predicts the merge via `git merge-tree --write-tree` (no mutation) and refuses only when the SHA both trails target AND conflicts. Remediation: merge target, resolve, re-attest.
 1. **CLEAR exists** — a `MergeClear` row is present for this PR, re-read from the DB (not an in-memory copy), all fields populated.
-2. **SHA still matches** — re-fetch the PR/MR's current HEAD SHA from the forge; it must equal `reviewed_sha`. If the branch has been force-pushed or new commits added, the CLEAR is stale. (Transport: `teatree.core.merge_execution` resolves the CLEAR's host kind from the ticket `issue_url` (`github.com` → GitHub, `gitlab*` → GitLab) and delegates the forge I/O to a `CodeHostBackend` built via `core.backend_registry` — the `gh`/`glab` argv lives on the backend impls; `merge_execution` keeps the verdict/transient/head-moved classification and error f-strings, so `core` never imports `teatree.backends`.)
+2. **SHA still matches** — re-fetch the PR/MR's current HEAD SHA from the forge; it must equal `reviewed_sha`. If the branch has been force-pushed or new commits added, the CLEAR is stale. (Transport: `teatree.core.merge` resolves the CLEAR's host kind from the ticket `issue_url` (`github.com` → GitHub, `gitlab*` → GitLab) and delegates the forge I/O to a `CodeHostBackend` built via `core.backend_registry` — the `gh`/`glab` argv lives on the backend impls; `merge.execution` keeps the verdict/transient/head-moved classification and error f-strings, so `core` never imports `teatree.backends`.)
 3. **CI still green** — re-evaluate against **the forge's live required-checks** for the target branch (GitHub's branch-protection list, or GitLab's head pipeline status — the authoritative set at merge time), not the `gh_verify_result` snapshot saved on the `MergeClear` row. All currently-required checks must pass at the moment of merge execution, not just at CLEAR issuance. (A required check may have been added after the CLEAR was issued; the live list is the source of truth.)
 4. **Not draft** — the PR must not be in draft state.
 5. **blast_class respected** — a `blast_class == substrate` CLEAR is draft-locked and needs a per-PR human sign-off (invariant 4), satisfied by EITHER (a) a per-CLEAR `human_authorizer` re-presented via `--human-authorized <owner-id>`, OR (b) the CLEAR's overlay standing at `autonomy = full` — the standing grant stands in for `--human-authorized`. Either way the agent runs the keystone (never raw `gh`). The carve-out removes ONLY the per-PR sign-off; the floor (cold-review, SHA-bind, required-checks, not-draft, never-lockout) is unchanged. `--human-authorized` never unlocks a non-substrate CLEAR; below `full` the `human_authorizer` stays mandatory.
