+++ src/teatree/core/management/commands/review.py
@@ -25,7 +25,7 @@ import typer
 from django_typer.management import TyperCommand, command, initialize
 
 from teatree.core.gates.schema_guard import SelfDbMigrationError, require_current_schema
-from teatree.core.merge_execution import fetch_live_head_sha, fetch_required_checks_status
+from teatree.core.merge import fetch_live_head_sha, fetch_required_checks_status
 from teatree.core.models import Finding, ReviewVerdict, ReviewVerdictError, Ticket
 from teatree.utils.url_slug import pr_ref_from_url
