+++ src/teatree/core/merge/ci_rollup.py
@@ -1,0 +1,224 @@
+"""Forge transport resolution + CI/pipeline verdict classification (GitHub + GitLab).
+
+The lowest layer of the ``core/merge`` package: ``_code_host_for`` resolves the
+merge-transport backend via ``core.backend_registry`` (core never imports
+``teatree.backends`` — the §17.6.2 ``core ↛ backends`` edge), and the three thin
+``fetch_*`` delegators plus the rollup/pipeline classifiers live here so that
+both ``pr_slug_resolution`` (which probes live head SHAs) and ``execution``
+(which re-checks CI at merge time) depend DOWN on this module — the §1993 cut
+that keeps the intra-package DAG acyclic under ``forbid_circular_dependencies``.
+"""
+
+import logging
+from typing import TYPE_CHECKING, TypedDict, cast
+
+from teatree.core.backend_protocols import PrMergeState, rollup_query_failed
+from teatree.core.backend_registry import get_backend_provider
+
+if TYPE_CHECKING:
+    from teatree.core.backend_protocols import CodeHostBackend
+
+logger = logging.getLogger(__name__)
+
+
+def _code_host_for(host_kind: str) -> "CodeHostBackend":
+    """The merge-transport backend for *host_kind*, resolved via the registry.
+
+    Core never imports ``teatree.backends`` (the §17.6.2 ``core ↛ backends``
+    edge); it reaches a built backend ONLY through
+    :func:`core.backend_registry.get_backend_provider`. The token/base_url are
+    left empty — the merge-RPC runners use ambient ``gh``/``glab`` auth, the
+    same as the former in-module ``_run_gh``/``_run_glab`` did. When the
+    backends app is not installed the provider is the fail-safe
+    ``_UnconfiguredProvider``, whose ``build_*`` RAISE a clear ``RuntimeError``
+    (loud-failure: a merge in an unconfigured context fails visibly rather than
+    silently shelling out).
+    """
+    provider = get_backend_provider()
+    if host_kind == "gitlab":
+        return provider.build_gitlab_host(token="", base_url="")
+    return provider.build_github_host(token="")
+
+
+def fetch_live_head_sha(slug: str, pr_id: int, *, host_kind: str = "github") -> str:
+    """The PR/MR's current head SHA from the forge (never a branch ref) — §17.4.3 step 2.
+
+    Delegates to the registry-resolved :class:`CodeHostBackend`
+    (:func:`_code_host_for`); the gh/glab argv lives in the backend.
+    """
+    return _code_host_for(host_kind).fetch_live_head_sha(slug=slug, pr_id=pr_id)
+
+
+def fetch_pr_merge_state(slug: str, pr_id: int, *, host_kind: str = "github") -> PrMergeState:
+    """Whether the PR/MR is already merged, and at which commit — §928 reconciliation.
+
+    A lost post-hook (process kill / DB lock / rollback between
+    :func:`execute_bound_merge` and :func:`record_merge_and_advance`)
+    leaves the PR merged on the forge while the CLEAR is still unconsumed
+    and the FSM has not advanced. The retry must detect "already merged
+    by us" and run the post hook idempotently rather than re-issuing the
+    irreversible merge (which both forges refuse — GitHub 405, GitLab 405
+    / 406 — a permanent brick) or failing the SHA precondition forever.
+    Returns an empty state on any forge error so the caller falls through to
+    the normal (fail-closed) precondition path. The backend normalises both
+    forges' state to the uppercase ``"MERGED"`` ``PrMergeState.is_merged`` reads.
+    """
+    return _code_host_for(host_kind).fetch_pr_merge_state(slug=slug, pr_id=pr_id)
+
+
+def fetch_pr_is_draft(slug: str, pr_id: int, *, host_kind: str = "github") -> bool:
+    """Whether the PR/MR is in draft state — §17.4.3 step 4.
+
+    Delegates to the registry-resolved :class:`CodeHostBackend`; GitLab reads
+    ``draft``/``work_in_progress`` and GitHub ``isDraft`` inside the backend.
+    """
+    return _code_host_for(host_kind).fetch_pr_is_draft(slug=slug, pr_id=pr_id)
+
+
+class _RollupEntry(TypedDict, total=False):
+    """One ``gh ... statusCheckRollup`` entry — CheckRun or StatusContext."""
+
+    conclusion: object
+    status: object
+    state: object
+
+
+def _classify_check(check: object) -> str:
+    """Map one rollup entry to ``green`` / ``pending`` / ``failed``.
+
+    CheckRun entries use ``conclusion`` + ``status``; legacy StatusContext
+    entries use ``state``. A non-dict entry is ignored by the caller.
+    """
+    if not isinstance(check, dict):
+        return ""
+    entry = cast("_RollupEntry", check)
+    conclusion = str(entry.get("conclusion") or "").upper()
+    status = str(entry.get("status") or "").upper()
+    state = str(entry.get("state") or "").upper()
+    if status and status != "COMPLETED":
+        return "pending"
+    if conclusion in {"SUCCESS", "NEUTRAL", "SKIPPED"} or state == "SUCCESS":
+        return "green"
+    if state == "PENDING":
+        return "pending"
+    return "failed"
+
+
+def _rollup_verdict(statuses: list[str]) -> str:
+    if "failed" in statuses:
+        return "failed"
+    if "pending" in statuses:
+        return "pending"
+    return "green"
+
+
+def fetch_required_checks_status(slug: str, pr_id: int, *, host_kind: str = "github") -> str:
+    """Live required-checks verdict for the PR/MR head — §17.4.3 step 3.
+
+    Evaluated against the forge's live rollup at merge time (the authoritative
+    set), NOT the ``gh_verify_result`` snapshot saved on the CLEAR. Returns
+    ``"green"`` only when every reported check concluded successfully;
+    ``"pending"`` while any is still running; otherwise the failing state.
+
+    The backend returns the RAW rollup (GitHub ``statusCheckRollup`` entries,
+    GitLab pipeline entries); core does the verdict classification here so the
+    §17.4.3 ``green``/``pending``/``failed`` semantics stay in one place. A
+    backend query failure surfaces as the :data:`ROLLUP_QUERY_FAILED` sentinel
+    → ``failed``; an empty rollup means no required checks → ``green``. GitLab
+    needs the head SHA to pick the right (non-merge-train) pipeline, fetched via
+    :func:`fetch_live_head_sha`.
+    """
+    backend = _code_host_for(host_kind)
+    rollup = backend.fetch_required_checks_rollup(slug=slug, pr_id=pr_id)
+    if rollup_query_failed(rollup):
+        return "failed"
+    if host_kind == "gitlab":
+        if not rollup:
+            return "green"
+        head_sha = backend.fetch_live_head_sha(slug=slug, pr_id=pr_id)
+        head = _select_gitlab_head_pipeline(list(rollup), head_sha, slug=slug, pr_id=pr_id)
+        if head is None:
+            return "failed"
+        return _classify_gitlab_pipeline(str(head.get("status") or ""))
+    statuses = [verdict for check in rollup if (verdict := _classify_check(check))]
+    return _rollup_verdict(statuses) if statuses else "green"
+
+
+_GITLAB_PIPELINE_GREEN_STATUSES = frozenset({"success", "manual", "skipped"})
+_GITLAB_PIPELINE_PENDING_STATUSES = frozenset(
+    {"pending", "running", "preparing", "scheduled", "waiting_for_resource", "created"},
+)
+
+
+def _classify_gitlab_pipeline(status: str) -> str:
+    """Map a GitLab pipeline status string to ``green`` / ``pending`` / ``failed``.
+
+    GitLab pipeline statuses (per the REST API documentation): ``created``,
+    ``waiting_for_resource``, ``preparing``, ``pending``, ``running``,
+    ``success``, ``failed``, ``canceled``, ``skipped``, ``manual``,
+    ``scheduled``. ``success`` / ``manual`` / ``skipped`` are green;
+    ``failed`` / ``canceled`` are failed; everything else is pending.
+    """
+    s = status.lower()
+    if s in _GITLAB_PIPELINE_GREEN_STATUSES:
+        return "green"
+    if s in _GITLAB_PIPELINE_PENDING_STATUSES:
+        return "pending"
+    return "failed"
+
+
+class _GitlabPipeline(TypedDict, total=False):
+    """One entry of ``glab api .../merge_requests/<iid>/pipelines``."""
+
+    id: object
+    sha: object
+    ref: object
+    source: object
+    status: object
+
+
+def _is_merge_train_pipeline(pipeline: _GitlabPipeline) -> bool:
+    ref = str(pipeline.get("ref") or "")
+    source = str(pipeline.get("source") or "")
+    return source == "merge_train" or "/train" in ref
+
+
+def _select_gitlab_head_pipeline(
+    pipelines: list[object],
+    head_sha: str,
+    *,
+    slug: str,
+    pr_id: int,
+) -> _GitlabPipeline | None:
+    """Pick the pipeline for the MR head commit, ignoring merge-train pipelines.
+
+    The ``…/merge_requests/<iid>/pipelines`` endpoint interleaves merge-train
+    pipelines (each on a transient train SHA, often canceled the moment the
+    train re-bases) ahead of the real head-branch pipeline, so ``pipelines[0]``
+    is not reliably the head pipeline. Match on the MR head SHA instead. When
+    the head SHA is known but no pipeline matches it, the head commit has no
+    pipeline of its own — return ``None`` so the caller fails closed rather
+    than reading an unrelated commit's pipeline. The newest non-train pipeline
+    is used only when the head SHA could not be fetched at all.
+    """
+    entries = [cast("_GitlabPipeline", p) for p in pipelines if isinstance(p, dict)]
+    candidates = [e for e in entries if not _is_merge_train_pipeline(e)]
+    if head_sha:
+        for pipeline in candidates:
+            if str(pipeline.get("sha") or "") == head_sha:
+                return pipeline
+        logger.info(
+            "merge_execution: no GitLab pipeline matches MR head %s for %s#%s "
+            "(non-train candidates: %s) — failing closed",
+            head_sha,
+            slug,
+            pr_id,
+            [str(p.get("sha") or "") for p in candidates],
+        )
+        return None
+    logger.info(
+        "merge_execution: GitLab MR head SHA unavailable for %s#%s — falling back to newest non-train pipeline",
+        slug,
+        pr_id,
+    )
+    return candidates[0] if candidates else None
