+++ tests/teatree_core/test_merge_anti_vacuity_gate.py
@@ -19,7 +19,7 @@ import pytest
 from django.test import TestCase
 
 from teatree.config import UserSettings
-from teatree.core.merge_execution import MergePreconditionError, merge_ticket_pr
+from teatree.core.merge import MergePreconditionError, merge_ticket_pr
 from teatree.core.models import MergeClear, Ticket
 from tests.teatree_core.test_merge_execution import _GhStub
